fix: validate and parse

+++ scripts/validate.js
@@ -8,7 +8,7 @@ const validateResultPath = path.join(__dirname, '..', `validate_result.md`)
 if (mdPath.endsWith('.md')) {
   validateMd(path.join(__dirname, '../docs', mdPath.replace(/^docs\//, '')))
 } else {
-  fs.unlinkSync(validateResultPath)
+  fs.existsSync(validateResultPath) && fs.unlinkSync(validateResultPath)
   const mdFileDir = path.join(__dirname, '../docs', mdPath)
   fs.readdir(mdFileDir, (err, files) => {
     files.forEach(fileName => {
@@ -18,20 +18,21 @@ if (mdPath.endsWith('.md')) {
 }
 
 function validateMd(mdPath, fileName) {
+  if(!mdPath.endsWith('.md')) return;
   fs.readFile(mdPath, { encoding: 'utf-8' }, (err, data) => {
     if (err) {
       throw err
     }
     const result = parse(data)
     if (!fileName) {
-      console.log(result.notTranslate)
+      console.log(result.notTranslated)
       return;
     }
 
     fs.writeFileSync(validateResultPath, `
-#### ${fileName} not Translate content 
+#### ${fileName} not translated content 
 \`\`\`
-${result.notTranslate.join('\n')}
+${result.notTranslated.join('\n')}
 \`\`\`
 ---
